DAOS-875 object: Extract enumeration packing code (2/2)
Move the extracted types and functions to dss. No change to the logic,
except for replacing enum_anchor_reset_hkey with daos_hash_set_zero.

Note that this adds an undesirable vos-dependency to dss due to
vos_iter_type_t and vos_iter_param_t. Alternatively, we could define new
types, but they would be largely equivalent to the two vos type.

Reviewed-on: https://review.hpdd.intel.com/32799
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
