+++ src/include/daos_types.h
@@ -92,11 +92,17 @@ typedef struct {
 #define DAOS_HASH_HKEY_LENGTH	24
 
 static inline void
+daos_hash_set_zero(daos_hash_out_t *hash_out)
+{
+	memset(&hash_out->body[DAOS_HASH_HKEY_START], 0,
+	       DAOS_HASH_HKEY_LENGTH);
+}
+
+static inline void
 daos_hash_set_eof(daos_hash_out_t *hash_out)
 {
 	memset(&hash_out->body[DAOS_HASH_HKEY_START], -1,
 	       DAOS_HASH_HKEY_LENGTH);
-
 }
 
 static inline bool
